toc: Open items when clicking on the label for a h2 item
When clicking the label for a h2 item, the onclick called
`selectTocListItem`, which opened the children. Then the URL got
updated, which again called `selectTocListItem`. But this time, the
item was already selected (due to the first call) and hence toggled
the children's visibility, effectively hiding them.

So we now force children open for hash changes.